requirements.txt
asgiref==3.7.2
bleach==6.0.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-admin-interface==0.26.0
django-allauth==0.54.0
django-ckeditor==6.7.0
django-colorfield==0.9.0
django-compression-middleware==0.5.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-markdownify==0.9.3
django-mdeditor==0.1.20
django-mptt==0.14.0
django-mptt-admin==2.5.1
django-pygments==0.5
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
Markdown==3.4.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tinycss2==1.1.1
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
zope.event==5.0
zope.interface==6.0
zstandard==0.21.0
